% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_ticks.R
\name{log_ticks}
\alias{log_ticks}
\title{Logarithmic tick marks}
\usage{
log_ticks(x, log = FALSE)
}
\arguments{
\item{x}{The values of the axis in question.}

\item{log}{Logical indicating if x is already logarithmic. Defaults to
\code{FALSE}.}
}
\value{
A list containing a vector each of the major and minor tick mark
  positions which is intended to be used in conjuncton with \code{axis}, see
  examples.
}
\description{
Calculate the position of major and minor ticks for a log10 axis.
}
\examples{
x <- 10^seq(1, 4, length.out = 100)
ticks <- log_ticks(x)
plot(x, stats::rnorm(100), log = "x", xaxt = "n")
axis(1, at = ticks$major_ticks)
axis(1, at = ticks$minor_ticks, labels = FALSE, tcl = -0.25)

# Use log = TRUE for input that is already logarithmic:
x2 <- log10(x)
ticks2 <- log_ticks(x2, log = TRUE)
}
