% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_range.R
\name{map_range}
\alias{map_range}
\title{map_range}
\usage{
map_range(x, in_start = min(x, na.rm = TRUE), in_end = max(x, na.rm =
  TRUE), out_start = 0, out_end = 1)
}
\arguments{
\item{x}{foo}

\item{in_start}{bar}

\item{in_end}{baz}

\item{out_start}{bla}

\item{out_end}{blubb}
}
\value{
blargh
}
\description{
map_range
}
\examples{
map_range(1:3)                  # 0.0, 0.5, 1.0
map_range(1:3, 1, 10)           # 0.00, 0.11, 0.22
map_range(1:3, 0, 10)           # 0.1, 0.2, 0.3
map_range(1:3, 0, 10, 0, 100)   # 10, 20, 30
map_range(1:3, 1, 10, 10, 100)  # 10, 20, 30
}
