% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/not_in.R
\name{not_in}
\alias{not_in}
\alias{\%!in\%}
\title{Negation of \%in\%}
\usage{
a \%!in\% b
}
\arguments{
\item{a}{The values to be matched.}

\item{b}{The values to be matched against.}
}
\value{
The negated output of \code{\link[base:match]{\%in\%}}.
}
\description{
Provides a "not in" like in Python that is in my opinion prettier than
\code{!x \%in\% y}.
}
\details{
\code{!x \%in\% y} works, too, because \code{\%in\%} takes precedence before
\code{!}, see \link[base]{Syntax}.
}
