% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round.R
\name{round_to_next_multiple}
\alias{round_to_next_multiple}
\alias{ceiling_to_next_multiple}
\alias{floor_to_next_multiple}
\title{Round towards a multiple of n}
\usage{
round_to_next_multiple(x, n)

ceiling_to_next_multiple(x, n)

floor_to_next_multiple(x, n)
}
\arguments{
\item{x}{The value to round.}

\item{n}{The value which determines the rounding target.}
}
\description{
Round x up or down to the nearest multiple of n.
}
\note{
Symmetrical rounding to the next power of ten can also be archived
  using \code{\link[base]{round}} with a negative \code{digits} argument.
}
\examples{
round_to_next_multiple(120, 50)  # 100
round_to_next_multiple(17, 5)  # 15
floor_to_next_multiple(195, 100)  # 100
ceiling_to_next_multiple(10.3, 2)  # 12

}
