% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/not_in.R
\name{not_in}
\alias{not_in}
\alias{\%!in\%}
\title{Negation of \%in\%}
\usage{
a \%!in\% b
}
\arguments{
\item{a}{The values to be matched.}

\item{b}{The values to be matched against.}
}
\value{
The negated output of \code{\link[base:match]{\%in\%}}.
}
\description{
Provides a "not in" like in Python by negating \code{\%in\%}.
}
\details{
In my opinion this is prettier and more readable than using \code{!x \%in\%
y}. By the way, this notation works without parentheses because \code{\%in\%}
takes precedence before \code{!}, see \link[base]{Syntax}.
}
\examples{
42 \%!in\% 1:10  # TRUE

}
